ch-ch-ch-ch-changes

+++ src/Book.java
@@ -6,9 +6,9 @@ public class Book
     private String genre;
     private String publisher;
     private int releaseYear;
-    private int isbn;
+    private long isbn;
 
-    public Book(String title, String author, String language, String genre, String publisher, int releaseYear, int isbn)
+    public Book(String title, String author, String language, String genre, String publisher, int releaseYear, long isbn)
     {
         this.title = title;
         this.author = author;
@@ -45,7 +45,7 @@ public class Book
         return releaseYear;
     }
 
-    public int getIsbn(){
+    public long getIsbn(){
         return isbn;
     }
 
@@ -79,7 +79,7 @@ public class Book
         this.releaseYear = releaseYear;
     }
 
-    public void setIsbn(int isbn)
+    public void setIsbn(long isbn)
     {
         this.isbn = isbn;
     }
